Move instruments into metric_instrument.go
